% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Synthetic data set to investigate three-class ROC data.}
\format{
A data frame with 100 rows and 10 variables (9 classifiers):
\describe{
  \item{trueClass}{A factor, indicating the class membership of the
  individuals.}
  \item{Class1, Class2, Class3, Class4, Class5, Class6, Class7,  Class8, Class9}{Measurements of a specific classifier obtained from the patients that underwent the
  clinical study.}
}
}
\usage{
cancer
}
\description{
A dataset containing randomly generated measurements from three diagnostic classes as they may
arise in a cancer investigation.
}
\keyword{datasets}
