% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinROC.R
\docType{package}
\name{trinROC-package}
\alias{trinROC}
\alias{trinROC-package}
\title{trinROC: Statistical Tests for Assessing Trinormal ROC Data}
\description{
Several statistical test functions as well as a function for exploratory data analysis to investigate classifiers allocating individuals to one of three disjoint and ordered classes. In a single classifier assessment the discriminatory power is compared to classification by chance. In a comparison of two classifiers the null hypothesis corresponds to equal discriminatory power of the two classifiers.
}
\details{
See \code{vignette("Overview", package = "trinROC")} for an overview of the package.
  Further, \code{sd()}, \code{var()} and \code{cov()} are chosen with \code{options(trinROC.MLE = TRUE)} according to the maximum likelihood estimates (\code{default}). Change to sample
  estimates by setting \code{options(trinROC.MLE = FALSE)}
}
\references{
Noll, S., Furrer, R., Reiser, B. and Nakas, C. T. (2019).
  Inference in ROC surface analysis via a trinormal model-based testing approach.
  \emph{Stat}, accepted.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.math.uzh.ch/reinhard.furrer/trinROC}
}

}
\author{
\strong{Maintainer}: Reinhard Furrer \email{reinhard.furrer@math.uzh.ch} [contributor]

Authors:
\itemize{
  \item Samuel Noll \email{uncle.sam@gmx.net}
}

Other contributors:
\itemize{
  \item Benjamin Reiser [contributor]
  \item Christos T. Nakas \email{cnakas@uth.gr} [contributor]
}

}
